requirements.txt
pandas
langextract
pdfplumber
python-dotenv
google-cloud-storage
requests
beautifulsoup4
